Add more acceptance specs (and run into trouble)
Add two specs, one to cover an empty grid (which passed anyway), and one
to cover a 'Block' pattern (which stays static). Unfortunately making
these tests pass is looking difficult!

This commit adds a fairly hefty if statement to our implementation, now
we're going to look for a different way.